+++ src/ProductPage.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import FrequentQuestions from './frequentQuestions/FrequentQuestions.jsx';
 import Overview from './overview/Overview';
 import Ratings from './ratings/Ratings';
-import RelatedItems from './relatedItems/RelatedItems';
+import RelatedItems from './relatedItems/RelatedItems.jsx';
 import Search from './Search.jsx';
 
 function ProductPage({ product }) {
