+++ src/relatedItems/styles.css
@@ -66,7 +66,7 @@
 .card_slide {
   position: relative;
   /* width: 100%; */
-  padding: 0 5%;
+  padding: 0 50px;
 }
